chart view use data get from server

+++ app/src/main/java/com/github/lzyzsd/androidstockchart/Point.java
@@ -4,10 +4,10 @@ package com.github.lzyzsd.androidstockchart;
  * Created by Bruce on 2/27/15.
  */
 public class Point {
-    public float x;
+    public long x;
     public float y;
 
-    public Point(float x, float y) {
+    public Point(long x, float y) {
         this.x = x;
         this.y = y;
     }
